refactor: update import paths from partners.const to partners.data
