+++ app/shared/components/blocks/our-partners/partners.data.ts
